refactor: extract creating user logic into controller
