+++ src/controller/__init__.py
@@ -1,3 +1,4 @@
 from .article import create_article_controller, BodyCreateArticle
 from .sig import create_sig_controller, BodyCreateSIG, update_sig_controller, BodyUpdateSIG
 from .pig import create_pig_controller, BodyCreatePIG, update_pig_controller, BodyUpdatePIG
+from .user import create_user_controller, BodyCreateUser
